fix Mac tests / update build agent to pass along compiled artifacts (#1595)
* fix Mac tests / update build agent

* (test_osn_input) add missing hotkeys for MAC_SCREEN_CAPTURE
* (nodeobs_settings.cpp) fix "&& within ||" Xcode warning

(test_nodeobs_settings) do not set ApplyServiceSettings to true on mac

* If this property is enabled on Mac, then the encoder settings will be overwritten
and cause the test to fail- because you're not allowed to change certain fields.

rename macos_avcapture to fix test

* avcapture plugins were renamed so they can be looked up directly in JS code need to update the tests- https://github.com/streamlabs/obs-studio/pull/679/commits/b53043d5bb8072ad9a07ae1109d2f988ed84fdc6

(Video.cpp - CreateVideo) add bounds check on input from JS

(test_osn_input) ignore ndi_source (exists on client but not on agent)
(deleteConfigFiles) check if folder exists first before trying to remove files
* fix issue where it would stop deleting config files on the mac-Crashpad folder.
now all config files are deleted properly and the ini files will be properly
reset before next test

(main.yml) update CI steps to pass along build artifacts
(controller.cpp) print out error if unable to spawn obs64 (helpful on CI build agent)

+++ tests/osn-tests/util/general.ts
@@ -38,7 +38,7 @@ let showHideInputHotkeys: string[] = ['SHOW_SCENE_ITEM.WASAPI_OUTPUT_CAPTURE', '
 'HIDE_SCENE_ITEM.AV_CAPTURE_INPUT_V2','SHOW_SCENE_ITEM.SPOUT_CAPTURE','HIDE_SCENE_ITEM.SPOUT_CAPTURE',
 'SHOW_SCENE_ITEM.SLIDESHOW_V2', 'HIDE_SCENE_ITEM.SLIDESHOW_V2', 'SHOW_SCENE_ITEM.TEXT_GDIPLUS_V3', 'HIDE_SCENE_ITEM.TEXT_GDIPLUS_V3',
 'SHOW_SCENE_ITEM.MACOS_AVCAPTURE', 'HIDE_SCENE_ITEM.MACOS_AVCAPTURE', 'SHOW_SCENE_ITEM.MACOS_AVCAPTURE_FAST', 'HIDE_SCENE_ITEM.MACOS_AVCAPTURE_FAST',
-'SHOW_SCENE_ITEM.SCK_AUDIO_CAPTURE', 'HIDE_SCENE_ITEM.SCK_AUDIO_CAPTURE'];
+'SHOW_SCENE_ITEM.SCK_AUDIO_CAPTURE', 'HIDE_SCENE_ITEM.SCK_AUDIO_CAPTURE', 'SHOW_SCENE_ITEM.MAC_SCREEN_CAPTURE', 'HIDE_SCENE_ITEM.MAC_SCREEN_CAPTURE'];
 
 export {showHideInputHotkeys};
 
@@ -76,22 +76,21 @@ export function deleteConfigFiles(): void {
     const fs = require('fs');
     const path = require('path');
     const configFolderPath = path.join(path.normalize(__dirname), '..', 'osnData/slobs-client');
-    let files;
-    let currentFile: string;
-
-    try {
-        files = fs.readdirSync(configFolderPath);
-
+    if (fs.existsSync(configFolderPath)) {
+        let currentFile: string;
+        let files = fs.readdirSync(configFolderPath);
         files.forEach(file => {
             if (file !== 'node-obs') {
                 currentFile = file;
-                fs.unlinkSync(path.join(configFolderPath, file));
+                try {
+                    fs.unlinkSync(path.join(configFolderPath, file));
+                } catch(error) {
+                    if (error.code === "EBUSY") {
+                        throw ('Error: the file ' + currentFile + ' or slobs-client folder is busy');
+                    }
+                }
             }
         });
-    } catch(error) {
-        if (error.code === "EBUSY") {
-            throw ('Error: the file ' + currentFile + ' or slobs-client folder is busy');
-        }
     }
 }
